fix #2

+++ src/main/java/vavi/util/tag/id3/v2/impl/ID3v2FrameV230.java
@@ -410,7 +410,7 @@ logger.warning("no key for: " + getID());
         }
         ret[9] = flag2;
 
-        short content_offset = 10; // first byte used for content
+        int content_offset = 10; // first byte used for content
 
         // decompressed size, if compressed
         if (compression) {
